% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{AIC.efit}
\alias{AIC.efit}
\title{AIC}
\usage{
\method{AIC}{efit}(obj)
}
\arguments{
\item{obj}{A fitted model object.}
}
\description{
extracts the Akaike Information Criterion from efit model objects.
}
