% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raneffects.R
\name{blup}
\alias{blup}
\alias{blup.raneffects}
\title{blup}
\usage{
blup(obj, ...)

\method{blup}{raneffects}(obj, ...)
}
\arguments{
\item{obj}{A \code{raneffects} object.
#'}
}
\value{
a \code{matrix} giving the marginal likelihood of N for each observation
}
\description{
\code{blup} estimates the best linear unbiased predictor of the latent abundance
or occupancy state using empirical Bayes methods. A port of the similar function
in \code{unmarked}.
}
