% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootN}
\alias{bootN}
\alias{bootN.efitMS}
\alias{bootN.efitMNO}
\title{bootN}
\usage{
bootN(obj, ...)

\method{bootN}{efitMS}(obj, B = 50, trim = 0.1, ...)

\method{bootN}{efitMNO}(obj, B = 50, trim = 0.1, ...)
}
\arguments{
\item{obj}{A fitted model object.}

\item{B}{The number of bootstrap replicates (default = 50).
There must by covariate values for every parameter in \code{obj}.}

\item{trim}{The fraction of bootstrap samples to be trimmed before computation of the mean
and standard errors. Samples are trimmed symmetrically from the ends of the
bootstrap distribution.}
}
\value{
a \code{data.frame} giving the predictions for each spatial unit
 as well as the overall abundance estimate for the region with associated
 SE and confidence intervals.
}
\description{
\code{bootN} estimates abundance from a fitted model for the region covered by
the sampled sites. Standard errors and confidence intervals are estimated
using a bootstrap procedure by resampling sites.
}
\examples{
 rem<- san_nic_open$removals
 emf <- eFrameMNO(y=rem, numPrimary=8)
 mod <- remMNO(~1, ~1, ~1, ~1, data=emf)
 Nhat<- bootN(mod)

}
