% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{calcN}
\alias{calcN}
\alias{calcN.efit}
\alias{calcN.efitM}
\title{calcN}
\usage{
calcN(obj, ...)

\method{calcN}{efit}(obj, covs, off.set = NULL, CI.level = 0.95, ...)

\method{calcN}{efitM}(obj, covs, off.set = NULL, CI.level = 0.95, ...)
}
\arguments{
\item{obj}{A fitted model object.}

\item{covs}{A \code{data.frame} of covariates for each spatial unit.
There must by covariates for every parameter in \code{obj}.
if no covariates are used then \code{covs} should be a data.frame
with one column indicating cell number.}

\item{off.set}{Either a scalar offset value to apply to each spatial unit
for prediction (e.g. cell area) or a vector of the same length as \code{nrow(covs)}.}
}
\value{
a \code{data.frame} giving the predictions for each spatial unit
 as well as the overall abundance estimate for the region with associated
 SE and confidence intervals.
}
\description{
\code{calcN} estimates abundance for a defined region from a fitted model.
The user must supply the number of spatial units (cells) or a data.frame
of covariate values for each spatial unit.
}
\examples{
 counts<- san_nic_pre$counts
 site.df<- san_nic_pre$traps
 emf <- eFrame(y=counts, siteCovs=site.df)
 mod <- nmix(~1, ~1, data=emf)
 Nhat<- calcN(mod, site.df)

}
