% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{coef}
\alias{coef}
\alias{coef.efit}
\title{coef}
\usage{
coef(obj, ...)

\method{coef}{efit}(obj, type, ...)
}
\arguments{
\item{obj}{A fitted model object.}
}
\description{
extracts coefficients for efit model objects.
}
