% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFrame.R
\name{eFrame}
\alias{eFrame}
\title{eFrame}
\usage{
eFrame(y, siteCovs = NULL)
}
\arguments{
\item{y}{An MxJ matrix of the observed measured data, where M is the
number of sites and J is the maximum number of observations per site.}

\item{SiteCovs}{A \code{data.frame} of covariates that vary at the
site level. This should have M rows and one column per covariate}
}
\value{
a \code{eFrame} holding data containing the response and
 covariates required for each model
}
\description{
\code{eFrame} creates an eFrame data object for use with Nmixture
or occupancy type models
}
\examples{
 counts<- san_nic_pre$counts
 emf <- eFrame(y=counts)
 summary(emf)

}
