% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFrame.R
\name{eFrameDS}
\alias{eFrameDS}
\title{eFrameDS}
\usage{
eFrameDS(distance, size, siteID, cutpoints, w, bin_nums = FALSE)
}
\arguments{
\item{distance}{A vector of distances (or distance bin number) for each
detected individual. Actual distances are the distance (m) to the midpoint
of each distance bin.}

\item{size}{the number of individuals recorded for each distance measurement
(i.e. groupsize).}

\item{siteID}{A vector indicating the camera ID for each distance measurement.}

\item{cutpoints}{vector of bin cutpoints indicating the distance to the end of
each bin. cutpoints should begin at zero and end with w.}

\item{w}{Truncation distance or maximum distance from camera that will be considered
in the analysis. All distances or bins further than this will be discarded.}

\item{bin_nums}{A logical indicating whether \code{distance} records
distances or bin numbers.  Bin numbers are assumed to be numbered from 1.}
}
\value{
a \code{eFrameDS} holding data containing the data suitable for estimating a
camera detection function
}
\description{
\code{eFrameDS} creates an eFrame data object for data collected by
distance sampling, specifically from remote cameras.  Distance sampling
is used to estimate the area of a camera viewshed by fitting a detection
function and estimating the effective sampled area.
}
\examples{
 counts<- san_nic_pre$counts
 emf <- eFrame(y=counts)
 summary(emf)

}
