% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFrame.R
\name{eFrameGP}
\alias{eFrameGP}
\title{eFrameGP}
\usage{
eFrameGP(catch, effort, index = NULL, ieffort = NULL)
}
\arguments{
\item{catch}{A vector of removals for each period.}

\item{effort}{A vector of removal effort employed during each period (i.e. trapnights).}

\item{index}{Optional vector of relative abundance indices for each period.}
}
\value{
a \code{eFrameGP} holding data suitable for use in \code{remGP}
}
\description{
\code{eFrameGP} creates an eFrameGP data object for use with the single site
removal estimator \code{remGP()}.
}
\examples{
 rem<- san_nic_rem$rem
 ym<- san_nic_rem$ym # detections from additional monitoring
 catch<- apply(rem,2,sum)
 effort<- rep(nrow(rem), length(catch))
 index<- apply(ym,2,sum)

 emf<-eFrameGP(catch, effort, index)
 summary(emf)

}
