% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFrame.R
\name{eFrameMNO}
\alias{eFrameMNO}
\title{eFrameMNO}
\usage{
eFrameMNO(
  y,
  numPrimary,
  siteCovs = NULL,
  primaryCovs = NULL,
  primaryPeriod = NULL
)
}
\arguments{
\item{y}{An MxJxT matrix of the observed removal data, where M is the
number of sites, J is the maximum number of removal (secondary) periods
per site and T is the number of primary periods.}

\item{numPrimary}{the number of primary periods. For each primary period, the
population is assumed to be closed.}

\item{primaryCovs}{A \code{data.frame} of covariates that vary at the
site x primary period (M x T) level.}

\item{primaryPeriod}{A MxT matrix of integers indicating the time gap between
primary periods for each site}

\item{SiteCovs}{A \code{data.frame} of covariates that vary at the
site level. This should have M rows and one column per covariate}
}
\value{
a \code{eFrameMNO} holding data containing the response and
 covariates required for open population removal models
}
\description{
\code{eFrameMNO} creates an eFrameMNO data object for use with open population
multinomial removal models using the robust design where sampling occurs over
a number of primary and secondary periods.
}
\examples{
 rem<- san_nic_rem$rem
 ym<- san_nic_rem$ym # detections from additional monitoring

 emf<-eFrameMNO(rem, numPrimary=1)
 summary(emf)

}
