% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFrame.R
\name{eFrameMNS}
\alias{eFrameMNS}
\title{eFrameMNS}
\usage{
eFrameMNS(y, numPrimary, siteCovs = NULL, obsCovs = NULL, delta = NULL)
}
\arguments{
\item{y}{An MxJxT matrix of the observed removal data, where M is the
number of sites, J is the maximum number of removal (secondary) periods
per site and T is the number of primary periods.}

\item{numPrimary}{the number of primary periods. For each primary period, the
population is assumed to be closed.}

\item{obsCovs}{A list of matrices or data.frames of variables varying within sites.
Each matrix or data.frame must be of dimension MxJ.}

\item{delta}{A vector with T elements giving the time units between primary periods for each site
beginning with 1 for the first primary period. A default value of 1 is used to
indicate equal time intervals between primary periods.}

\item{SiteCovs}{A \code{data.frame} of covariates that vary at the
site level. This should have M rows and one column per covariate}
}
\value{
a \code{eFrameMNS} holding stacked data for each primary period.
}
\description{
\code{eFrameMNS} creates an eFrameMNS 'stacked' data object for use with closed population
multinomial removal models using the robust design where sampling occurs over
a number of primary and secondary periods. Data for each primary period is 'stacked'
into rows with an indicator variable added to identify each primary period. The data can
then be analysed using closed population removal models to estimate the trend in abundance
between primary periods.
}
\examples{
 rem<- san_nic_rem$rem
 ym<- san_nic_rem$ym # detections from additional monitoring

 emf<-eFrameMNO(rem, numPrimary=1)
 summary(emf)

}
