% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFrame.R
\name{eFrameMS}
\alias{eFrameMS}
\title{eFrameMS}
\usage{
eFrameMS(y, obsPerSeason, siteCovs = NULL, obsCovs = NULL)
}
\arguments{
\item{y}{An MxSJ matrix of the observed detection/non-detection data
where M is the number of sites, S is the number of seasons and J is
the maximum number of observations per season.}

\item{obsPerSeason}{scalar of the maximum number of observations (replicates)
per season.}

\item{siteCovs}{A \code{data.frame} of covariates that vary at the
site level. This should have M rows and one column per covariate}

\item{obsCovs}{A list of matrices or data.frames of variables varying within sites.
Each matrix or data.frame must be of dimension MxJ.}
}
\value{
a \code{eFrameMS} holding data containing the response and
 covariates required for \code{occuMS}
}
\description{
\code{eFrameMS} creates an eFrameMS data object for use with multi-season
occupancy models \code{occMS} where sampling occurs over a number of
primary and secondary periods.
}
\examples{
### NEED TO SUPPLY GOOD EXAMPLE DATA
 rem<- san_nic_rem
 #'
 emf<-eFrameMS(y=rem, obsPerSeason=XXX, type="removal")
 summary(emf)

}
