% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFrame.R
\name{eFrameR}
\alias{eFrameR}
\title{eFrameR}
\usage{
eFrameR(y, type, siteCovs = NULL)
}
\arguments{
\item{y}{An MxJ matrix of the observed removal data, where M is the
number of sites and J is the maximum number of removal (primary)
periods per site. Each primary period can consist of k secondary
periods but this is not used here.}

\item{type}{sampling method. Currently supports "removal" for removal
sampling and "double" for double observer sampling}

\item{SiteCovs}{A \code{data.frame} of covariates that vary at the
site level. This should have M rows and one column per covariate}
}
\value{
a \code{eFrameR} holding data containing the response and
 covariates required for removal models
}
\description{
\code{eFrameR} creates an eFrameR data object for use with removal models.
}
\examples{
 rem<- san_nic_rem$rem
 emf<- eFrameR(rem, type="removal")
 summary(emf)

}
