% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFrame.R
\name{eFrameREST}
\alias{eFrameREST}
\title{eFrameREST}
\usage{
eFrameREST(y, stay, cens, area, active_hours, siteCovs = NULL)
}
\arguments{
\item{y}{An MxJ matrix of the observed number of encounters at a camera over
a specified period (usually 24h but could be less - see \code{active_hours})
where M is the number of sites and J is the maximum number of observations per site.}

\item{stay}{A vector of the observed staying times, which are the difference
between the exit and entry times for all (or a sample) of individuals}

\item{cens}{A vector indicating whether both entry and exit times were
observed for each of the staying times (\code{cens}=1) or only the entry time
was observed with the exit time unknown (\code{cens}=0).  In the latter case,
the staying time is the maximum time that the individual was observed to be in
front of the camera.}

\item{area}{A scalar indicating The area of the camera viewshed in square meters.
This area is assumed to have a detection probability of 1.0 (effective sampling area).}

\item{active_hours}{A scalar indicating the number of hours in each 24 hour
period that the species is active}
}
\value{
a \code{eFrameREST} holding data containing the response and
 covariates required for the REST model
}
\description{
\code{eFrameREST} creates an eFrame data object for use with the REST
model applied to camera trap data using the MLE given in:
Nakashima,Y., Fukasawa, K., and Samejima H. (2018). J.App.Ecol,55,735-744.
}
