% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFrame.R
\name{eFrameRGPM}
\alias{eFrameRGPM}
\title{eFrameRGPM}
\usage{
eFrameRGPM(y, ym, numPrimary, siteCovs = NULL, primaryCovs = NULL, type)
}
\arguments{
\item{y}{An NxJ matrix of the additional monitoring data, where N is the
number of monitored sites and J is the maximum number of primary periods
per site as for \code{y}.}

\item{cells}{vector indictaing which of the \code{y} sites were subject
to monitoring}

\item{Z}{integer indicating the number of secondary periods per
primary period for \code{y1}}
}
\value{
a \code{eFrameRGP} holding data containing the response and
 covariates required for removal models
}
\description{
\code{eFrameRGP} creates an eFrameRGP data object for use with generalized removal
models using the robust design where sampling occurs over a number of primary and
secondary periods.
}
\examples{
 rem<- san_nic_rem$rem
 y1<- san_nic_rem$y1 # detections from additional monitoring
 mtraps<- san_nic_rem$cells
 nights<- san_nic_rem$nights

 emf<-eFrameRM(rem, y1, mtraps, nights, type="removal")
 summary(emf)

}
