% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFrame.R
\name{eFrameRM}
\alias{eFrameRM}
\title{eFrameRM}
\usage{
eFrameRM(y, ym, Z, type, siteCovs = NULL)
}
\arguments{
\item{y}{An MxJ matrix of the observed removal data, where M is the
number of sites and J is the maximum number of removal (primary)
periods per site. Each primary period can consist of k secondary
periods but this is not used here.}

\item{ym}{An MxJ matrix of the additional monitoring data,}

\item{Z}{integer indicating the number of secondary periods per
primary period for \code{ym}}

\item{type}{sampling method. Currently supports "removal" for removal
sampling and "double" for double observer sampling}
}
\value{
a \code{eFrameRM} holding data containing the response and
 covariates required for removal models
}
\description{
\code{eFrameRM} creates an eFrameRM data object for use with removal models
that include additional monitoring at some or all of the sites.
}
\examples{
 rem<- san_nic_rem$rem
 ym<- san_nic_rem$ym # detections from additional monitoring
 nights<- san_nic_rem$nights

 emf<-eFrameRM(rem, ym, nights, type="removal")
 summary(emf)

}
