% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFrame.R
\name{eFrameS}
\alias{eFrameS}
\title{eFrameS}
\usage{
eFrameS(ys, siteCovs = NULL, obsCovs = NULL, delta = NULL)
}
\arguments{
\item{ys}{A \code{data.frame} of the observed data for each site \code{M}
in rows and secondary periods \code{J} in columns, indexed by session (primary period).
The data.frame \code{ys} must contain a column \code{session} with at least two unique values.}

\item{siteCovs}{A \code{data.frame} of covariates that vary at the
site level. This should have M rows and one column per covariate}

\item{obsCovs}{A list of matrices or data.frames of variables varying within sites.
Each matrix or data.frame must be of dimension MxJ.}

\item{delta}{A vector with elements giving the time units between primary periods for each site
beginning with 1 for the first primary period. A default value of 1 is used to
indicate equal time intervals between primary periods.}
}
\value{
a \code{eFrameS} holding stacked data for each primary period.
}
\description{
\code{eFrameS} creates an eFrameS 'stacked' data object for use with closed population
n-mixture or single season occupancy models using the robust design where sampling occurs over
a number of primary and secondary periods. Data for each primary period is 'stacked'
into rows with an indicator variable added to identify each primary period. The data can
then be analysed using closed population models to estimate the trend in abundance
between primary periods.
}
\examples{
 ys<- san_nic_open$counts

 emf<-eFrameS(ys)
 summary(emf)

}
