% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{fitted.efitMNO}
\alias{fitted.efitMNO}
\title{fitted.efitMNO}
\usage{
\method{fitted}{efitMNO}(obj, na.rm = FALSE)
}
\arguments{
\item{obj}{A fitted model object.}
}
\description{
calculates fitted values for the open population
multinomial removal model.
}
