% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{fitted.efitR}
\alias{fitted.efitR}
\title{fitted}
\usage{
fitted.efitR(obj, na.rm = TRUE)
}
\arguments{
\item{obj}{A fitted model object.}
}
\description{
calculates the fitted values for a efit model.
}
