% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_encounters}
\alias{make_encounters}
\title{make_encounters}
\usage{
make_encounters(sites, events)
}
\arguments{
\item{sites}{A \code{data.frame} of sites (camera locations). The first three columns MUST
be in the following order. 1. Site (or camera) ID; 2. first day of operation for each camera
(either day of year or date); 3. Last day of operation for each camera.}

\item{events}{A \code{data.frame} of encounters for each camera, The columns MUST be in
the following order. 1. Site (or camera) ID; 2. Day of year (or date) of each encounter;
3. the number of individuals encountered (i.e. group size).}
}
\value{
a \code{matrix} with the number of encounters for each camera with dimensions
M x J where M is the number of sites and J is the maximum number of days operation.  Cameras
operating for less than J days are assigned NA for those days.
}
\description{
\code{make_encounters} is a convenience function to make an encounter matrix y suitable for
use in \code{eFrameREST}. It converts long data to the wide format with one row per site.
}
\examples{
sites<- HogDeer$sites
encounters<- HogDeer$encounters
y<- make_encounters(sites, encounters)

}
