% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmix.R
\name{nmix}
\alias{nmix}
\title{nmix}
\usage{
nmix(lamformula, detformula, data, K, mixture=c("P", "NB", "ZIP"),
starts, method="BFGS", se=TRUE, ...)
}
\arguments{
\item{lamformula}{formula for the latent abundance component.}

\item{detformula}{formula for the detection component.  Only
site-level covariates are allowed for the detection component.
This differs from the similar model in \code{unmarked}.}

\item{data}{A \code{eFrame} object containing the response (counts)
and site-level covariates. see \code{\link{eFrame}} for how to format
the required data.}

\item{K}{Integer upper index of integration for abundance. This should be
set high enough so that it does not affect the parameter estimates. Note
that computation time will increase with K}

\item{mixture}{Distribution modelfor the latent abundance, either Poisson (P)
or Negative-binomial (NB).}

\item{starts}{Initial values for parameters}

\item{method}{Optimisation method}

\item{se}{flag to return the standard error (hessian).}
}
\value{
a \code{efit} model object.
}
\description{
\code{nmix} fits the N-mixture model of Royle et al (2004)
}
\examples{
 emf <- eFrame(y=counts, siteCovs=site.df)
 mod <- nmix(~1, ~1, data=emf)
 Nhat<- calcN(mod, ncells=55)

}
