% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmixS.R
\name{nmixS}
\alias{nmixS}
\title{nmixS}
\usage{
nmixS(lamformula, detformula, data, K, mixture=c("P", "NB"),
starts, method="BFGS", se=TRUE, ...)
}
\arguments{
\item{lamformula}{formula for the latent abundance component.}

\item{detformula}{formula for the detection component.  Only
site-level covariates are allowed for the detection component.
This differs from the similar model in \code{unmarked}.}

\item{data}{A \code{eFrameS} object containing the response (counts)
for each site, indexed by primary period (sessions) and site-level covariates.
see \code{\link{eFrameS}} for how to format the required data.}

\item{K}{Integer upper index of integration for abundance. This should be
set high enough so that it does not affect the parameter estimates. Note
that computation time will increase with K}

\item{mixture}{Distribution model for the latent abundance, either Poisson (P)
or Negative-binomial (NB).}

\item{starts}{Initial values for parameters}

\item{method}{Optimisation method}

\item{se}{flag to return the standard error (hessian).}
}
\value{
a \code{efit} model object.
}
\description{
\code{nmixS} fits the N-mixture model of Royle et al (2004) to 'stacked' data
from \code{M} sites over \code{T} primary periods (sessions) with each primary period
consisting of \code{J} secondary periods. Currently supported models include
the Poisson and Negative binomial
}
\examples{
 counts<- san_nic_open$counts
 emf <- eFrameS(y=counts)
 mod <- nmixS(~.season, ~1, data=emf)
 Nhat<- calcN(mod)

}
