% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occM.R
\name{occM}
\alias{occM}
\title{occuM}
\usage{
occuM(lamformula, detformula, data, knownocc, starts, method="BFGS", se=TRUE, ...)
}
\arguments{
\item{lamformula}{formula for the latent occupancy component.}

\item{detformula}{formula for the detection component.  Only
site-level covariates are allowed for the detection component.
This differs from the similar model in \code{unmarked}.}

\item{data}{A \code{eFrame} object containing the response (either (0/1) or counts)
and site-level covariates. see \code{\link{eFrame}} for how to format
the required data. count data will get trunctated to (0/1) by \code{occuM()}.}

\item{starts}{Initial values for parameters}

\item{method}{Optimisation method}

\item{se}{flag to return the standard error (hessian).}

\item{knownocc}{Vector of row numbers of sites that are known to be occupied.}
}
\value{
a \code{efit} model object.
}
\description{
\code{occuM} fits the occupancy model of McKenzie et al. (2002).
}
\examples{
 counts<- san_nic_pre$counts
 emf <- eFrame(y=counts)
 mod <- occuM(~1, ~1, data=emf)
 Nhat<- calcN(mod)

}
