% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occMS.R
\name{occMS}
\alias{occMS}
\title{occMS}
\usage{
occMS(lamformula, gamformula, epsformula, detformula,
data, starts, method="BFGS", se=TRUE, ...)
}
\arguments{
\item{lamformula}{formula for the latent (initial) occupancy component.}

\item{gamformula}{formula for the latent colonisation component. If season
specific colonisation/extinction parameters are desired, then use the
reserved keyword \code{.season}, (i.e. \code{~.season}).}

\item{epsformula}{formula for the latent survival component.}

\item{detformula}{formula for the detection component.  Only
site- or \code{.season}-level covariates are allowed for
the detection component. This differs from the similar model in \code{unmarked}.}

\item{data}{A \code{eFrameMS} object containing the detection/non-detection
data (0/1) and site-level covariates. see \code{\link{eFrameMS}} for how to format
 the required data. count data will get trunctated to (0/1) by \code{occMS()}.}

\item{starts}{Initial values for parameters}

\item{method}{Optimisation method}

\item{se}{flag to return the standard error (hessian).}
}
\value{
a \code{efit} model object.
}
\description{
\code{occMS} fits the dynamic multi-season occupancy model of MacKenzie et al. (2003).
This is a port of the \code{colext} function in \code{unmarked}.
}
\examples{
 counts<- san_nic_open$index
 emf <- eFrameMS(df=counts)
 mod <- occMS(~1, ~1, ~.season, ~1, data=emf)
 Nhat<- calcN(mod)

}
