% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occRN.R
\name{occRN}
\alias{occRN}
\title{occuRN}
\usage{
occuRN(lamformula, detformula, data, K, starts, method="BFGS", se=TRUE, ...)
}
\arguments{
\item{lamformula}{formula for the latent abundance component.}

\item{detformula}{formula for the detection component.  Only
site-level covariates are allowed for the detection component.
This differs from the similar model in \code{unmarked}.}

\item{data}{A \code{eFrame} object containing the response (either (0/1) or counts)
and site-level covariates. see \code{\link{eFrame}} for how to format
the required data. count data will get trunctated to (0/1) by \code{occuRN()}.}

\item{K}{Integer upper index of integration for abundance. This should be
set high enough so that it does not affect the parameter estimates. Note
that computation time will increase with K}

\item{starts}{Initial values for parameters}

\item{method}{Optimisation method}

\item{se}{flag to return the standard error (hessian).}
}
\value{
a \code{efit} model object.
}
\description{
\code{occuRN} fits the occupancy/abundance model of Royle & Nichols (2003).
}
\examples{
 counts<- san_nic_pre$counts
 emf <- eFrame(y=counts)
 mod <- occuRN(~1, ~1, data=emf)
 Nhat<- calcN(mod)

}
