% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occRNS.R
\name{occRNS}
\alias{occRNS}
\title{occuRNS}
\usage{
occuRNS(lamformula, detformula, data, K, starts, method="BFGS", se=TRUE, ...)
}
\arguments{
\item{lamformula}{formula for the latent abundance component.}

\item{detformula}{formula for the detection component.  Only
site-level covariates are allowed for the detection component.
This differs from the similar model in \code{unmarked}.}

\item{data}{A \code{eFrameS} object containing the response for each site,
indexed by primary period (sessions) and site-level covariates.see
 \code{\link{eFrameS}} for how to format the required data. count data
 will get trunctated to (0/1) by \code{occuRNS()}.}

\item{K}{Integer upper index of integration for abundance. This should be
set high enough so that it does not affect the parameter estimates. Note
that computation time will increase with K}

\item{starts}{Initial values for parameters}

\item{method}{Optimisation method}

\item{se}{flag to return the standard error (hessian).}
}
\value{
a \code{efitS} model object.
}
\description{
\code{occuRNS} fits the occupancy/abundance model of Royle & Nichols (2003) to 'stacked' data
from \code{M} sites over \code{T} primary periods (sessions) with each primary period
consisting of \code{J} secondary periods. Currently supported model for the abundance
process is the Poisson
}
\examples{
 counts<- san_nic_open$index
 emf <- eFrameS(counts)
 mod <- occuRNS(~.season, ~1, data=emf)  # counts truncated to (0/1)
 Nhat<- calcN(mod)

}
