% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{occTraject}
\alias{occTraject}
\alias{occTraject.efitMS}
\title{occTraject}
\usage{
occTraject(obj, ...)

\method{occTraject}{efitMS}(obj, type = c("projected", "smoothed"), mean = TRUE, ...)
}
\arguments{
\item{obj}{A fitted model object.}
}
\description{
extracts occupancy trajectories from occuMS objects.
}
