% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occuM.R
\name{occuM}
\alias{occuM}
\title{occuM}
\usage{
occuM(lamformula, detformula, data, knownocc, starts, method="BFGS", se=TRUE, ...)
}
\arguments{
\item{lamformula}{formula for the latent abundance component.}

\item{detformula}{formula for the detection component.  Only
site-level covariates are allowed for the detection component.
This differs from the similar model in \code{unmarked}.}

\item{data}{A \code{eFrame} object containing the response (counts)
and site-level covariates. see \code{\link{eFrame}} for how to format
the required data.}

\item{starts}{Initial values for parameters}

\item{method}{Optimisation method}

\item{se}{flag to return the standard error (hessian).}

\item{knownocc}{Vector of row numbers of sites that are known to be occupied.}
}
\value{
a \code{efit} model object.
}
\description{
\code{occuM} fits the occupancy model of McKenzie et al. (2002).
}
\examples{
 emf <- eFrame(y=counts, siteCovs=site.df)
 mod <- occuM(~1, ~1, data=emf)
 Nhat<- calcN(mod, ncells=55)

}
