% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occuRN.R
\name{occuRN}
\alias{occuRN}
\title{occuRN}
\usage{
occuRN(lamformula, detformula, data, K, starts, method="BFGS", se=TRUE, ...)
}
\arguments{
\item{lamformula}{formula for the latent abundance component.}

\item{detformula}{formula for the detection component.  Only
site-level covariates are allowed for the detection component.
This differs from the similar model in \code{unmarked}.}

\item{data}{A \code{eFrame} object containing the response (counts)
and site-level covariates. see \code{\link{eFrame}} for how to format
the required data.}

\item{K}{Integer upper index of integration for abundance. This should be
set high enough so that it does not affect the parameter estimates. Note
that computation time will increase with K}

\item{starts}{Initial values for parameters}

\item{method}{Optimisation method}

\item{se}{flag to return the standard error (hessian).}
}
\value{
a \code{efit} model object.
}
\description{
\code{occuRN} fits the occupancy/abundance model of Royle & Nichols (2003).
}
\examples{
 emf <- eFrame(y=counts, siteCovs=site.df)
 mod <- occuRN(~1, ~1, data=emf)
 Nhat<- calcN(mod, ncells=55)

}
