% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raneffects.R
\name{postSamples}
\alias{postSamples}
\alias{postSamples.raneffects}
\title{postSamples}
\usage{
postSamples(obj, ...)

\method{postSamples}{raneffects}(obj, nsims = 100, ...)
}
\arguments{
\item{obj}{An object of class 'raneffects'.}
}
\value{
a \code{matrix} giving the posterior samples of N
}
\description{
\code{postSamples} performs estimation of the posterior distribution of N using
empirical Bayes methods. A port of the similar function in \code{unmarked}.
}
