% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFrame.R
\name{print.eFrame}
\alias{print.eFrame}
\title{print.eFrame}
\usage{
\method{print}{eFrame}(object, ...)
}
\arguments{
\item{object}{An \code{eFrame} object.}
}
\value{
a \code{list} containing various summaries of the data
}
\description{
\code{print} method for eFrame objects. Basically the same as \code{summary.eFrame}
}
\examples{
## uses san_nic_pre
 emf <- eFrame(y=counts)
 summary(emf)

}
