% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFrame.R
\name{print.eFrameGRM}
\alias{print.eFrameGRM}
\title{print.eFrameGRM}
\usage{
\method{print}{eFrameGRM}(object, ...)
}
\arguments{
\item{object}{An \code{eFrameRM} object.}
}
\value{
a \code{list} containing various summaries of the data
}
\description{
\code{print} method for eFrameRM objects. Basically the same as \code{summary.eFrame}
}
\examples{
 ## uses san_nic_rem
 emf <- eFrameRM(y=rem, y1=y1, cells=cells, Z=nights)
 summary(emf)

}
