% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFrame.R
\name{print.eFrameMNO}
\alias{print.eFrameMNO}
\title{print.eFrameMNO}
\usage{
\method{print}{eFrameMNO}(object, ...)
}
\arguments{
\item{object}{An \code{eFrameMNO} object.}
}
\value{
a \code{list} containing various summaries of the data
}
\description{
\code{print} method for eFrameMNO objects. Basically the same as \code{summary.eFrame}
}
\examples{
 ## uses san_nic_rem
 emf <- eFrameMNO(y=rem, numPrimary=1)
 summary(emf)

}
