% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFrame.R
\name{print.eFrameMNS}
\alias{print.eFrameMNS}
\title{print.eFrameMNS}
\usage{
\method{print}{eFrameMNS}(object, ...)
}
\arguments{
\item{object}{An \code{eFrameMNS} object.}
}
\value{
a \code{list} containing various summaries of the data
}
\description{
\code{print} method for eFrameMNS objects. Basically the same as \code{summary.eFrame}
}
\examples{
 rem<- san_nic_open$removals
 emf <- eFrameMNS(rem)
 print(emf)

}
