% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFrame.R
\name{print.eFrameR}
\alias{print.eFrameR}
\title{print.eFrameR}
\usage{
\method{print}{eFrameR}(object, ...)
}
\arguments{
\item{object}{An \code{eFrameR} object.}
}
\value{
a \code{list} containing various summaries of the data
}
\description{
\code{print} method for eFrameR objects. Basically the same as \code{summary.eFrame}
}
\examples{
## uses san_nic_rem
 emf <- eFrameR(y=rem, type="removal")
 summary(emf)

}
