% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFrame.R
\name{print.eFrameREST}
\alias{print.eFrameREST}
\title{print.eFrameREST}
\usage{
\method{print}{eFrameREST}(object, ...)
}
\arguments{
\item{object}{An \code{eFrameREST} object.}
}
\value{
a \code{list} containing various summaries of the data
}
\description{
\code{print} method for eFrameREST objects. Basically the same as \code{summary.eFrame}
}
\examples{
 ## uses rest
 emf <- eFrameREST(y, stay, cens, area, active_hours)
 summary(emf)

}
