% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{profileCI}
\alias{profileCI}
\title{profileCI}
\usage{
profileCI(object, type, level = 0.95)
}
\arguments{
\item{type}{parameter type (i.e. "state","det")}

\item{level}{confidence level}

\item{obj}{A fitted model object.}
}
\description{
extracts profile likelihood confidence intervals for parameters
on the link scale.
}
