% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remCE.R
\name{remCE}
\alias{remCE}
\title{remCE}
\usage{
remCE(catch, effort, nboot)
}
\arguments{
\item{catch}{the number of removals recorded for each period.}

\item{effort}{the overall effort expended in each period (i.e. trapnights)}

\item{method}{the depletion method. Variable Effort Models: l= Leslie estimator,
ml= maximum likelihood estimator of Gould and Pollock (1997),
hosc = sampling coverage estimator for homogeneous model of Chao and Chang (1999),
hesc = sampling coverage estimator for heterogeneous model of Chao and Chang (1999), and
hemqle = maximum quasi likelihood estimator for heterogeneous model of Chao and Chang (1999).}

\item{nboot}{the number of bootstrap resamples for estimation of standard errors in the ml,
hosc, hesc and hemqle methods.}
}
\value{
a \code{efit} model object.
}
\description{
\code{remCE} fits various catch-effort models to removal data from a single site
(or combined data from many sites).  At least 3 removal periods
are required to fit these models.  A port of the \code{deplet} function
in the \code{fishmethods} package.
}
\examples{
 remCE(catch=y, effort=eff)

}
