% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remGP.R
\name{remGP}
\alias{remGP}
\title{remGP}
\usage{
remGP(catch, effort, index, starts, Nmax, se=TRUE, ...)
}
\arguments{
\item{catch}{the number of removals recorded for each period.}

\item{effort}{the overall effort expended in each period (i.e. trapnights)}

\item{index}{Index (i.e. count) data collected just before each removal period.
Index data can be any relative index of abundance.}

\item{starts}{Initial values for parameters}

\item{Nmax}{Maximum abundance considered in the optimisation}

\item{se}{flag to return the standard error (hessian).}
}
\value{
a \code{efit} model object.
}
\description{
\code{remGP} fits the catch-effort model of Gould & Pollock (1997) to removal
data from a single site (or combined data from many sites).  At least 3 removal periods
are required to fit this model.  Additionally, the model also accepts index (count)
data collected contemporaneously with the removal data.
}
\examples{
 rem<- san_nic_rem$rem
 emf <- eFrameR(y=rem)
 mod <- remPois(~1, ~1, data=emf)
 Nhat<- calcN(mod)

}
