% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remGR.R
\name{remGR}
\alias{remGR}
\title{remGR}
\usage{
remGR(lamformula, phiformula, detformula, data, mixture = c("P", "NB"), K,
                  starts, method="BFGS", se=TRUE, ...)
}
\arguments{
\item{lamformula}{formula for the latent abundance component.}

\item{phiformula}{formula for availability}

\item{detformula}{formula for the removal detection component.  Only
site-level covariates are allowed for the removal detection component.
This differs from the similar model in \code{unmarked}.}

\item{data}{A \code{eFrameR} object containing the response (counts)
and site-level covariates. see \code{\link{eFrameR}} for how to format
the required data.}

\item{K}{upper bound for superpopulation abundance}

\item{starts}{Initial values for parameters}

\item{method}{Optimisation method}

\item{se}{flag to return the standard error (hessian).}

\item{model}{for abundance, either Poisson 'P' or negative binomial 'NB'}
}
\value{
a \code{efit} model object.
}
\description{
\code{remGR} fits the generalized removal model to data collected from
repeated removal episodes from M sites over T primary periods with each primary consisting of J
secondary periods.
}
\examples{
 rem<- san_nic_rem$rem
 emf <- eFrameR(y=rem)
 mod <- remPois(~1, ~1, data=emf)
 Nhat<- calcN(mod)

}
