% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remGRMS.R
\name{remGRMS}
\alias{remGRMS}
\title{remGRMS}
\usage{
remGRMS(lamformula, detformula, mdetformula, data, mixture = c("P", "NB"), K,
                  starts, method="BFGS", se=TRUE, ...)
}
\arguments{
\item{lamformula}{formula for the latent abundance component.}

\item{detformula}{formula for the removal detection component.}

\item{mdetformula}{formula for the index detection component.}

\item{data}{A \code{eFrameGRM} object containing the response (counts, index)
and site-level covariates. see \code{\link{eFrameGRM}} for how to format
the required data.}

\item{K}{upper bound for superpopulation abundance}

\item{starts}{Initial values for parameters}

\item{method}{Optimisation method}

\item{se}{flag to return the standard error (hessian).}

\item{model}{for abundance, either Poisson 'P' or negative binomial 'NB'}
}
\value{
a \code{efitGRMS} model object.
}
\description{
\code{remGRMS} fits the generalized removal model to 'stacked' data collected from
repeated removal episodes from M sites over T primary periods with each primary consisting of J
secondary periods. The model also facilitates the analysis of index (count) data collected
in conjunction with the removal data to make joint inference on abundance.  Currently
supported models include the Poisson and Negative Binomial
}
\examples{
 rem<- san_nic_open$removal
 ym<- san_nic_open$index
 emf <- eFrameGRMS(rem, ym)
 mod <- remGRMS(~1, ~1, ~1, data=emf)
 Nhat<- calcN(mod)

}
