% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remMN.R
\name{remMN}
\alias{remMN}
\title{remMN}
\usage{
remPois(lamformula, detformula, data, starts, method="BFGS", se=TRUE, ...)
}
\arguments{
\item{lamformula}{formula for the latent abundance component.}

\item{detformula}{formula for the removal detection component.  Only
site-level covariates are allowed for the removal detection component.
This differs from the similar model in \code{unmarked}.}

\item{data}{A \code{eFrameR} object containing the response (counts)
and site-level covariates. see \code{\link{eFrameR}} for how to format
the required data.}

\item{mixture}{model for the latent abundance, either Poisson (\code{P}), negative binomial
(\code{NB}) or zero-inflated Poisson (\code{ZIP}).}

\item{starts}{Initial values for parameters}

\item{method}{Optimisation method}

\item{se}{flag to return the standard error (hessian).}
}
\value{
a \code{efit} model object.
}
\description{
\code{remMN} fits multinomial removal models (e.g. Haines 2019, Dorazio et al 2005)
to data from a number of primary periods where individuals removed are recorded for each site.
Currently supported models include the Poisson, Negative binomial and
zero-inflated Poisson (ZIP).
}
\examples{
 rem<- san_nic_rem$rem
 emf <- eFrameR(y=rem)
 mod <- remMN(~1, ~1, data=emf)
 Nhat<- calcN(mod)

}
