% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remMNS.R
\name{remMNS}
\alias{remMNS}
\title{remMNS}
\usage{
remMNS(lamformula, detformula, data, starts, method="BFGS", se=TRUE, ...)
}
\arguments{
\item{lamformula}{formula for the latent abundance component. For stacked data representing
multiple primary periods, the reserved terms \code{.season} or \code{.trend} can be added
to the formula to model abundance for each primary period as a factor (\code{.season}) or
numeric trend (\code{.trend})}

\item{detformula}{formula for the removal detection component.  Only
site-level covariates are allowed for the removal detection component.
This differs from the similar model in \code{unmarked}.}

\item{data}{A \code{eFrameR} object containing the response (counts)
and site-level covariates. see \code{\link{eFrameR}} for how to format
the required data.}

\item{mixture}{model for the latent abundance, either Poisson (\code{P}), negative binomial
(\code{NB}) or zero-inflated Poisson (\code{ZIP}).}

\item{starts}{Initial values for parameters}

\item{method}{Optimisation method}

\item{se}{flag to return the standard error (hessian).}
}
\value{
a \code{efit} model object.
}
\description{
\code{remMNS} fits multinomial removal models (e.g. Haines 2019, Dorazio et al 2005)
to 'stacked' data from from M sites over T primary periods with each primary consisting of J
secondary periods. Currently supported models include the Poisson, Negative binomial and
zero-inflated Poisson (ZIP).
}
\examples{
 rem<- san_nic_rem$rem
 emf <- eFrameR(y=rem)
 mod <- remMN(~1, ~1, data=emf)
 Nhat<- calcN(mod)

}
