% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remPois.R
\name{remPois}
\alias{remPois}
\title{remPois}
\usage{
remPois(lamformula, detformula, data, starts, method="BFGS", se=TRUE, ...)
}
\arguments{
\item{lamformula}{formula for the latent abundance component.}

\item{detformula}{formula for the removal detection component.  Only
site-level covariates are allowed for the removal detection component.
This differs from the similar model in \code{unmarked}. Currently
an intercept-only model is assumed for the occupancy component.}

\item{data}{A \code{eFrameR} object containing the response (counts)
and site-level covariates. see \code{\link{eFrameR}} for how to format
the required data.}

\item{starts}{Initial values for parameters}

\item{method}{Optimisation method}

\item{se}{flag to return the standard error (hessian).}
}
\value{
a \code{efit} model object.
}
\description{
\code{remPois} fits the multinomial-Poisson removal model to data from
a number of primary periods where individuals removed are recorded for each site.
}
\examples{
 data(snc)
 emf <- eFrameR(y=removed, siteCovs=site.df)
 mod <- remPois(~1, ~1, data=emf)
 Nhat<- calcN(mod, ncells=55)

}
