% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remPoisM.R
\name{remPoisM}
\alias{remPoisM}
\title{remPoisM}
\usage{
remPoisM(lamformula, detformula, data, K, starts, method="BFGS", se=TRUE, ...)
}
\arguments{
\item{lamformula}{formula for the latent abundance component.}

\item{detformula}{formula for the removal detection component.  Only
site-level covariates are allowed for the removal detection component.
This differs from the similar model in \code{unmarked}. Currently
an intercept-only model is assumed for the detection component of the occupancy data.}

\item{data}{A \code{eFrameRM} object containing the response (counts)
and site-level covariates. see \code{\link{eFrameRM}} for how to format
the required data.}

\item{K}{Integer upper index of integration for abundance. This should be
set high enough so that it does not affect the parameter estimates. Note
that computation time will increase with K.}

\item{starts}{Initial values for parameters}

\item{method}{Optimisation method}

\item{se}{flag to return the standard error (hessian).}
}
\value{
a \code{efit} model object.
}
\description{
\code{remPoisM} fits the combined removal model + royle-nichols
occupancy/abundance model to two sources of monitoring data. The first
(main) source of data is removal type data consisting of a number of primary
periods where individuals removed are recorded for each site. The second
source of data consists of standard occupancy data which is collected at the same
set of sites (or a proportion thereof) using a different method (e.g. cameras).
The model then estimates abundance in light of detections of individuals arising from both
sources of data.
}
\examples{
 ## uses san_nic_rem
 emf <- eFrameRM(y=rem, y1=y1, cells=cells, Z=nights)
 mod <- remPoisM(~1, ~1, data=emf)
 Nhat<- calcN(mod)

}
