% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{residuals.efitR}
\alias{residuals.efitR}
\title{residuals}
\usage{
\method{residuals}{efitR}(obj)
}
\arguments{
\item{obj}{A fitted model object.}
}
\description{
calculates working residuals from an efit model object.
}
