% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REST.R
\name{REST}
\alias{REST}
\title{REST}
\usage{
REST(formula, data, starts, method="BFGS", se=TRUE, ...)
}
\arguments{
\item{formula}{formula for animal density.}

\item{data}{A \code{eFrameREST} object containing the number of encounters
for each camera and the staying times. see \code{\link{eFrameREST}} for how to format
 the required data.}

\item{starts}{Initial values for parameters}

\item{method}{Optimisation method}

\item{se}{flag to return the standard error (hessian).}
}
\value{
a \code{efit} model object.
}
\description{
\code{REST} fits the random encounter and staying time model of
Nakashima et al (2018), which models the relationship between camera
encounter rate, staying time and density. The model assume that cameras
sample a viewshed of area \code{A} with perfect detection.
}
\examples{
 data(snc)
 emf <- eFrameREST(y=counts, stay=staying.times, cens= censor, area=A,
        active_hours=24, siteCovs=site.df)
 mod <- REST(~1, data=emf)
 Nhat<- calcN(mod, ncells=55)

}
