% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{san_nic_habitat.tif}
\alias{san_nic_habitat.tif}
\title{Simulated habitat (relative risk) on San Nicolas Island.}
\description{
A raster file at 100 meter resolution with simulated habitat values (either
\code{0} or \code{1}). Can be resampled for use with any of the models.
}
\examples{
 library(raster)
 habitat<- raster(system.file("extdata", "san_nic_habitat.tif", package="eradicate"))

}
