% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{san_nic_pre}
\alias{san_nic_pre}
\title{San_nic_pre}
\format{A list consisting of 2 elements
 \describe{
 \item{counts}{A matrix of counts with one row for each camera}
 \item{traps}{a data.frame of camera coordinates}
 }}
\usage{
san_nic_pre
}
\description{
Cat monitoring on San Nicolas Island
}
\details{
A simulated example dataset of pre-implementation monitoring of a cat population on San
Nicolas Island, California.  The data consist of the counts of cats recorded on 55
systematically spaced cameras on the island set for 30 nights.
}
\keyword{datasets}
