% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{san_nic_region}
\alias{san_nic_region}
\title{San Nicholas island.}
\description{
ESRI shapefile of San Nicolas Island.
}
\examples{
 library(sf)
 region<- read_sf(system.file("extdata", "shape/san_nic_region.shp", package="eradicate"))
}
