% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{san_nic_rest}
\alias{san_nic_rest}
\title{san_nic_rest}
\format{A list consisting of 2 elements
 \describe{
 \item{y}{An \code{M} x \code{J} matrix of counts of individuals detected at each camera
 where M is the number of cameras and J is the maximum number of days the camera was
 operating}
 \item{stay}{A vector of the observed staying times, which are the difference between
 the exit and entry times for all (or a sample) of individuals}
 \item{cens}{A vector indicating whether both entry and exit times were
 observed for each of the staying times (\code{cens}=1) or only the entry time
 was observed with the exit time unknown (\code{cens}=0).  In the latter case,
 the staying time is the maximum time that the individual was observed to be in
 front of the camera}
 \item{area}{The area of the camera viewshed (km2)}
 \item{active_hours}{The number of hours in a single 24 hour period that individuals of
 the target species were active for}
 \item{traps}{a data.frame with \code{M} rows of trap locations}
 }}
\usage{
san_nic_rest
}
\description{
Example camera monitoring data collected for use in the REST model
}
\details{
A simulated example dataset of camera detections collected for use with the
REST model of (Nakashima et al. 2018).  The data consist of a the number of individuals
detected at each camera as well as the time taken for each individual to cross the
camera viewshed (staying time).  It is assumed that detection within the camera
viewshed is perfect.
}
\keyword{datasets}
