% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{SE}
\alias{SE}
\alias{SE.efit}
\title{SE}
\usage{
SE(obj, ...)

\method{SE}{efit}(obj, type, ...)
}
\arguments{
\item{obj}{A fitted model object.}
}
\description{
extracts standard errors for efit model objects.
}
