% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{summary.efit}
\alias{summary.efit}
\alias{summary.efitREST}
\title{summary.efit}
\usage{
\method{summary}{efit}(object, ...)

\method{summary}{efitREST}(object, ...)
}
\arguments{
\item{obj}{A fitted model object.}
}
\value{
a \code{data.frame} giving the estimated parameters for the
 model in \code{obj} (on the link scale) and associated SE and 95% CI.
}
\description{
\code{summary} summarises an efit object.
}
\section{Functions}{
\itemize{
\item \code{summary.efitREST}: 
}}

\examples{
 emf <- eFrame(y=counts, siteCovs=site.df)
 mod <- nmix(~1, ~1, data=emf)
 summary(emf)

}
