% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{summary.efitGP}
\alias{summary.efitGP}
\title{summary.efitGP}
\usage{
\method{summary}{efitGP}(object, ...)
}
\arguments{
\item{obj}{A fitted model object.}
}
\value{
a \code{data.frame}
}
\description{
\code{summary} summarises an efitGP object giving the estimated parameters for N
 and catchability in \code{obj} (on the link scale) with associated SE and confidence
 intervals.
}
\examples{
 emf <- eFrame(y=counts, siteCovs=site.df)
 mod <- nmix(~1, ~1, data=emf)
 summary(emf)

}
