% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{summary.efitGPlist}
\alias{summary.efitGPlist}
\title{summary.efitGPlist}
\usage{
\method{summary}{efitGPlist}(object, ...)
}
\arguments{
\item{object}{A list of efitGP objects.}
}
\value{
a \code{list}
}
\description{
\code{summary} summarises an efitGPlist object giving the estimated parameters for the
 model in \code{obj} (on the link scale) with associated SE and confidence
 intervals.
}
\examples{
 emf <- eFrameGP(removals, effort, session=session)
 mod <- remGP(emf)
 summary(emf)

}
