% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFrame.R
\name{summary.eFrame}
\alias{summary.eFrame}
\title{summary.eFrame}
\usage{
\method{summary}{eFrame}(object, ...)
}
\arguments{
\item{object}{An \code{eFrame} object.}
}
\value{
a \code{list} containing various summaries of the data
}
\description{
\code{summary} method for eFrame objects.
}
\examples{
## uses san_nic_pre
 emf <- eFrame(y=counts)
 summary(emf)

}
