% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFrame.R
\name{summary.eFrameMNO}
\alias{summary.eFrameMNO}
\title{summary.eFrameGRM}
\usage{
\method{summary}{eFrameMNO}(object, ...)
}
\arguments{
\item{object}{An \code{eFrameMNO} object.}
}
\value{
a \code{list} containing various summaries of the data
}
\description{
\code{summary} method for eFrameMNO objects.
}
\examples{
 ## uses san_nic_rem
 emf <- eFrameMNO(y=rem, numPrimary=1)
 summary(emf)

}
