% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFrame.R
\name{summary.eFrameMNS}
\alias{summary.eFrameMNS}
\title{summary.eFrameMNS}
\usage{
\method{summary}{eFrameMNS}(object, ...)
}
\arguments{
\item{object}{An \code{eFrameMNS} object.}
}
\value{
a \code{list} containing various summaries of the data
}
\description{
\code{summary} method for eFrameMNS objects.
}
\examples{
 rem<- san_nic_open$removals
 emf <- eFrameMNS(rem)
 summary(emf)

}
