% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFrame.R
\name{summary.eFrameR}
\alias{summary.eFrameR}
\title{summary.eFrameR}
\usage{
\method{summary}{eFrameR}(object, ...)
}
\arguments{
\item{object}{An \code{eFrameR} object.}
}
\value{
a \code{list} containing various summaries of the data
}
\description{
\code{summary} method for eFrameR objects.
}
\examples{
## uses san_nic_rem
 emf <- eFrame(y=rem, type="removal")
 summary(emf)

}
