% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{traject}
\alias{traject}
\alias{traject.efitMS}
\alias{traject.efitMNO}
\title{traject}
\usage{
traject(obj, ...)

\method{traject}{efitMS}(obj, mean = TRUE, ...)

\method{traject}{efitMNO}(obj, na.rm = FALSE)
}
\arguments{
\item{obj}{A fitted model object.}
}
\description{
extracts trajectories for open population models
\code{occuMS} and \code{remMNO} objects.
}
