% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{update.efitMNO}
\alias{update.efitMNO}
\title{update}
\usage{
\method{update}{efitMNO}(
  obj,
  lamformula.,
  detformula.,
  gamformula.,
  omformula.,
  ...,
  evaluate = TRUE
)
}
\arguments{
\item{obj}{A fitted model object.}
}
\description{
updates efit model objects.
}
