% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{vcmat}
\alias{vcmat}
\alias{vcmat.efit}
\title{vcmat}
\usage{
vcmat(obj, ...)

\method{vcmat}{efit}(obj, type, ...)
}
\arguments{
\item{obj}{A fitted model object.}
}
\description{
extracts variance covaraiance matrix for efit model objects.
}
