% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{vcov.efit}
\alias{vcov.efit}
\title{vcov}
\usage{
vcov.efit(obj, type, ...)
}
\arguments{
\item{obj}{A fitted model object.}
}
\description{
extracts variance covariance matrix from efit model objects.
}
