% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raneffects.R
\name{bup}
\alias{bup}
\alias{bup.raneffects}
\title{bup}
\usage{
bup(obj, ...)

\method{bup}{raneffects}(obj, ...)
}
\arguments{
\item{obj}{A \code{raneffects} object.
#'}
}
\value{
a \code{matrix} giving the marginal likelihood of N for each observation
}
\description{
\code{bup} estimates the best unbiased predictor of the latent abundance
or occupancy state using empirical Bayes methods. A port of the similar function
in \code{unmarked}.
}
