% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{calcN}
\alias{calcN}
\alias{calcN.default}
\title{calcN}
\usage{
calcN(obj, ...)

\method{calcN}{default}(obj, ncells = NULL, covs = NULL, CI.level = 0.95, ...)
}
\arguments{
\item{obj}{A fitted model object.}

\item{covs}{A \code{data.frame} of covariates for each spatial unit.
There must by covariates for every parameter in \code{obj}.
Either one of \code{cells} or \code{covs} must be supplied.}

\item{cells}{The number of spatial units used for prediction}
}
\value{
a \code{data.frame} giving the overall abundance estimate for
 the region with associated SE and confidence intervals.
}
\description{
\code{calcN} estimates abundance for a defined region from a fitted model.
The user must supply the number of spatial units (cells) or a data.frame
of covariate values for each spatial unit.
}
\examples{
 emf <- eFrame(y=counts, siteCovs=site.df)
 mod <- nmix(~1, ~1, data=emf)
 Nhat<- calcN(mod, ncells=55)

}
