% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{calcP}
\alias{calcP}
\alias{calcP.efitR}
\alias{calcP.efitGP}
\alias{calcP.efitMNO}
\alias{calcP.efitMS}
\alias{calcP.efitMNS}
\title{calcP}
\usage{
calcP(obj, ...)

\method{calcP}{efitR}(obj, na.rm = TRUE)

\method{calcP}{efitGP}(obj, na.rm = TRUE)

\method{calcP}{efitMNO}(obj, na.rm = TRUE)

\method{calcP}{efitMS}(obj, na.rm = TRUE)

\method{calcP}{efitMNS}(obj, na.rm = TRUE)
}
\arguments{
\item{obj}{A fitted model object.}
}
\description{
calculates detection probability from a fitted model.
}
