% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_func.R
\name{dist_func}
\alias{dist_func}
\title{dist_func}
\usage{
dist_func(data, theta, keyfun = c("hn","hr") ...)
}
\arguments{
\item{data}{\code{eFrameDS} object containing the distance data and bin information.}

\item{formula}{formula for the scale parameter. Default is ~1.}

\item{theta}{angle of the camera viewshed in degrees}

\item{keyfun}{detection functions to fit (currently only half-normal ("hn") or
hazard rate ("hr")). No adjustments are currently permitted.}
}
\value{
a \code{efitDS} model object.
}
\description{
\code{dist_func} estimates a distance function from distance data collected from
camera traps.  The distance function is then used to estimate the effective area
of the camera viewshed by integration. Data are assumed to be in bins specified in
 \code{cutpoints} (see \code{eFrameDS} for more details).  The detection function is
 estimated using the function \code{ds()} from package \code{Distance} using point
 distance sampling. Currently no adjustments are included in the fitting.
}
\examples{
 ddata<- HogDeer$ddata
 cutpoints<- HogDeer$cutpoints
 emf<- eFrameDS(ddata$distance, ddata$size, ddata$camID, cutpoints, w=12.5, bin_num=TRUE)
 mods <- dist_func(emf, theta=HogDeer$theta)

}
