% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFrame.R
\name{print.eFrameGRMS}
\alias{print.eFrameGRMS}
\title{print.eFrameGRMS}
\usage{
\method{print}{eFrameGRMS}(object, ...)
}
\arguments{
\item{object}{An \code{eFrameGRMS} object.}
}
\value{
a \code{list} containing various summaries of the data
}
\description{
\code{print} method for eFrameGRMS objects. Basically the same as \code{summary.eFrame}
}
\examples{
 rem<- san_nic_open$removals
 idx<- san_nic_open$index
 emf <- eFrameGRMS(rem, idx)
 print(emf)

}
