% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raneffects.R
\name{raneffects}
\alias{raneffects}
\alias{raneffects.efitR}
\title{raneffects}
\usage{
raneffects(obj, ...)

\method{raneffects}{efitR}(obj, K, ...)
}
\arguments{
\item{obj}{A fitted model object.}

\item{K}{maximum value of \code{N} to integrate over. The default rarely makes sense}
}
\value{
a \code{matrix} giving the marginal likelihood of N for each observation
}
\description{
\code{raneffects} performs estimation of the marginal likelihood of N using
empirical Bayes methods.
}
