% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remGP.R
\name{remGP}
\alias{remGP}
\title{remGP}
\usage{
remGP(catch, effort, index, starts, Nmax, se=TRUE, ...)
}
\arguments{
\item{starts}{Initial values for parameters}

\item{Nmax}{Maximum search increment for abundance considered in the optimisation}

\item{se}{flag to return the standard error (hessian).}

\item{alpha}{quantile for (1-alpha) level confidence intervals}

\item{catch}{the number of removals recorded for each period.}

\item{effort}{the overall effort expended in each period (i.e. trapnights)}

\item{index}{Index (i.e. count) data collected in conjunction with the removal
data.  The index data is assumed to be collected just before each removal period and
can consist of any relative index of abundance.}
}
\value{
a \code{efit} model object.
}
\description{
\code{remGP} fits the catch-effort model of Gould & Pollock (1997) to removal
data from a single site (or combined data from many sites).  At least 3 removal periods
are required to fit this model.  Additionally, the model also accepts index (count)
data collected in conjunction with the removal data.
}
\examples{
 rem<- san_nic_rem$rem
 emf <- eFrameR(y=rem)
 mod <- remPois(~1, ~1, data=emf)
 Nhat<- calcN(mod)

}
