% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remGP.R
\name{remGP}
\alias{remGP}
\title{remGP}
\usage{
remGP(data, starts, K, method = "BFGS", se=TRUE, ...)
}
\arguments{
\item{data}{\code{eFrameGP} object containing the number removal/index data.}

\item{starts}{Initial values for parameters}

\item{K}{Integer representing upper bound for abundance for discrete integration}

\item{method}{optimsation method (see \code{?optim} for details)}

\item{se}{flag to return the standard error (hessian).}
}
\value{
a \code{efit} model object.
}
\description{
\code{remGP} fits the catch-effort model of Gould & Pollock (1997) to removal
data from a single site (or combined data from many sites).  At least 3 removal periods
are required to fit this model.  Additionally, the model also accepts index (count)
data collected in conjunction with the removal data.
}
\examples{
 rem<- san_nic_rem$rem
 emf <- eFrameR(y=rem)
 mod <- remPois(~1, ~1, data=emf)
 Nhat<- calcN(mod)

}
