% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remMNO.R
\name{remMNO}
\alias{remMNO}
\title{remMNO}
\usage{
remMNO(lamformula, gamformula, omformula, detformula, data, mixture = c("P", "NB","ZIP"),
               K, dynamics = c("constant", "autoreg", "notrend", "trend", "ricker", "gompertz"),
               fix = c("none", "gamma", "omega"), immigration=FALSE, iotaformula = ~1,
               starts, method="BFGS", se=TRUE, ...)
}
\arguments{
\item{lamformula}{formula for the latent abundance component.}

\item{gamformula}{formula for the latent recruitment process for parameter `gamma`.}

\item{omformula}{formula for the latent survival process for parameter `omega`.}

\item{detformula}{formula for the removal detection component.  Only
site-level covariates are allowed for the removal detection component.
This differs from the similar model in \code{unmarked}.}

\item{data}{A \code{eFrameMNO} object containing the removal data for each primary
period and site-level covariates. see \code{\link{eFrameMNO}} for how to format
 the required data.}

\item{mixture}{for abundance, either Poisson 'P', negative binomial 'NB' or zero-inflated
poisson 'ZIP'.}

\item{K}{upper bound for superpopulation abundance}

\item{dynamics}{Character string describing the type of population dynamics. `constant` indicates
that there is no relationship between omega and gamma. `autoreg` is an auto-regressive model in
which recruitment is modeled as `gamma*N[i,t-1]`, and `trend` is a model for exponential growth,
`N[i,t] = N[i,t-1]*gamma`, where `gamma` in this case represents the finite rate of increase.}

\item{fix}{If "`omega`", `omega` is fixed at 1. If "`gamma`", `gamma` is fixed at 0.}

\item{immigration}{Logical specifying whether immigration is included in the model}

\item{iotaformula}{formula for the number of immigrants per site, per time step.}

\item{starts}{Initial values for parameters}

\item{method}{Optimisation method}

\item{se}{flag to return the standard error (hessian).}
}
\value{
a \code{efit} model object.
}
\description{
\code{remMNO} fits the multinomial, open population removal model to data collected from
repeated removal episodes from M sites over T primary periods with each primary consisting of J
secondary periods. This is a port of the similar function in \code{unmarked}.
}
\examples{
 rem<- san_nic_open$removals
 emf <- eFrameMNO(y=rem, numPrimary=8)
 mod <- remMNO(~1, ~1, ~1, ~1, data=emf)
 Nhat<- calcN(mod)

}
