% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REST.R
\name{REST}
\alias{REST}
\title{REST}
\usage{
REST(formula, data, starts, method="BFGS", se=TRUE, ...)
}
\arguments{
\item{data}{A \code{eFrameREST} object containing the number of encounters
for each camera and the staying times. see \code{\link{eFrameREST}} for how to format
 the required data.}

\item{starts}{Initial values for parameters}

\item{method}{Optimisation method}

\item{se}{flag to return the standard error (hessian).}

\item{formula}{formula for animal density.}
}
\value{
a \code{efit} model object.
}
\description{
\code{REST} fits the random encounter and staying time model of
Nakashima et al (2018), which models the relationship between camera
encounter rate, staying time and density. The model assume that cameras
sample a viewshed of area \code{A} with perfect detection.
}
\examples{
 y<- san_nic_rest$y
 stay<- san_nic_rest$stay
 cens<- san_nic_rest$cens
 area<- san_nic_rest$area
 active_hours<- san_nic_rest$active_hours

 emf <- eFrameREST(y, stay, cens, area, active_hours)
 mod <- REST(~1, data=emf)
 Nhat<- calcN(mod)

}
