% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{se}
\alias{se}
\alias{se.efit}
\title{se}
\usage{
se(obj, ...)

\method{se}{efit}(obj, type, ...)
}
\arguments{
\item{obj}{A fitted model object.}
}
\description{
extracts standard errors for efit model objects.
}
