% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFrame.R
\name{summary.eFrameGRMS}
\alias{summary.eFrameGRMS}
\title{summary.eFrameGRMS}
\usage{
\method{summary}{eFrameGRMS}(object, ...)
}
\arguments{
\item{object}{An \code{eFrameGRMS} object.}
}
\value{
a \code{list} containing various summaries of the data
}
\description{
\code{summary} method for eFrameGRMS objects.
}
\examples{
 rem<- san_nic_open$removals
 idx<- san_nic_open$index
 emf <- eFrameGRMS(rem, idx)
 summary(emf)

}
